refactor(dashboard+signatur): shared ActiveImpactsList, dashboard mirrors signatur
Extracts planet-influence rendering into a new shared component
src/components/shared/ActiveImpactsList.tsx with two variants:
- full (Signatur page) keeps the "Warum?" toggle
- compact (Dashboard) drops the toggle, tighter spacing

TransitResonancePanels is now a thin adapter; Dashboard's hero
mounts <ActiveImpactsList variant="compact" maxItems={4} />, fed by
birthSign (client-side computeTodayPlanetInfluences).

PlanetCard in DailyChartHero removed; activePlanets prop dropped;
birthSign prop added. Dashboard.tsx passes birthSign instead of
impactPlanets. 7 obsolete PlanetCard tests migrated to the new
active-impacts-list suite.

Phase 4 of docs/plans/2026-04-20-dashboard-signatur-gaps.md.
